Refactoring codes put strategy in seperate folders
