+++ comps/agent/langchain/src/strategy/agentic_rag/prompt.py
@@ -1,0 +1,3 @@
+from langchain_core.prompts import ChatPromptTemplate
+
+rlm_rag_prompt = ChatPromptTemplate.from_messages([("human", "You are an assistant for question-answering tasks. Use the following pieces of retrieved context to answer the question. If you don't know the answer, just say that you don't know. Use three sentences maximum and keep the answer concise.Question: {question} Context: {context} Answer:"),])
